Prettier standard (#225)
* Add prettier-standard

* Format with prettier-standard

* Update babel-eslint

* Fix peer-dependency warnings

* Lint npm script

* Add husky and lint-staged

* Add server.js to lint

* package.json

Standard style alone wasn't quite working right with the switch to create-react-app and some of the es-next way of writing things. In order to get that working it requires fiddling with parser config/versions/setup.

At the same time I've been spoiled in other projects using prettier to enforce an opinionated format.

This introduces prettier-standard and adds it lint on commit to enforce a consistent style in line with standard style.