Query editor redesign (#425)
Redesign query editor experience

This branch got away from me. It started with making the schema sidebar toggle-able and became the branch for exploring a lot of UI changes I wanted to make.

Here's a quick summary of the big stuff:

* The new side nav/app nav bar is gone. Instead everything has been fit into a single toolbar at top of the query editor page
* The query editor page is now essentially the only page. all other management pages are drawers that slide out. Managing users, connections, selecting a query... navigating to a separate page for this felt jarring for me at times and I wanted to keep the user in the editor at all times.
* Schema sidebar is toggle-able
* Query chart is moved into main view - it is no longer in a separate tab. This makes the vis smaller, but I think it's a nice change. Before modifying a query and seeing the new results was a pain.

There is still more polish needed for this work! But it is functional and ready to be iterated on in smaller chunks.