Use v4 UUID instead of v1 for query cacheKey
This uses a purely random v4 UUID instead of a time-based v1 UUID for
the query cacheKey (which were initially introduced in
84aae812da4bfdcb30e5ce7103f2f8b9a2afd1af).

v1 UUID are based on current time and the hardware MAC address of the
machine where they are being generated (although the implementation in
the npm uuid module uses generates a random fake MAC address). As such
they have much more complex semantics than v4 UUIDs which are simply
randomly generated.

Unless there's a specific requirement for the special semantics of v1
UUIDs it is simpler and less error prone to simply go for v4 UUIDs
whenever just a unique identifier is needed.