Enhance support for Amazon Athena (#1048)
This change extends the driver data model to include a flag indicating
that the driver implements a method to cancel queries.
The client renders a "cancel query" button, disabled when the query is
not running, if the driver has this flag set. Cancellation is important
to minimize costs for queries that were issued by mistake

It extends the schema query to show a "partition" hint next to the data
type of the column. The query is compatible with Athena engine v1 and
v2. This is important as it hints users which columns to include in
their queries to minimize the amount of data scanned (i.e. predicate
pushdown)

Resolves #1008

Co-authored-by: Serafin Sedano <serafins@amazon.com>